<Refactor> : Separation of Ceoncerns wrt design
